+++ docs/src/hooks/useDot.ts
@@ -1,26 +1,32 @@
 import { useEffect, useState } from 'react';
 import doT from 'dot';
+import { Tps } from '@site/types/templates';
 import * as utils from 'templates-mo/lib/templates/utils';
 
 doT.templateSettings.strip = false;
-doT.templateSettings.varname = 'tps';
 
 interface Props {
-  templateName?: string;
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  tps?: Record<string, any>;
+  templateString?: string;
+  tps?: Partial<Tps>;
 }
 
-type templateFn = (obj: Record<string, any>) => string;
+type templateFn = (obj: Tps) => string;
 
-export const useDot = ({ templateString, tps = {} }) => {
+const DEFAULT_TPS: Tps = {
+  name: 'App',
+  answers: {},
+  utils,
+  u: utils,
+};
+
+export const useDot = ({ templateString, tps = {} }: Props) => {
   const [output, setOutput] = useState<string>(null);
 
   useEffect(() => {
     let result;
     try {
       const dotTemplate: templateFn = doT.template(templateString);
-      result = dotTemplate({ ...tps, utils, u: utils });
+      result = dotTemplate({ ...DEFAULT_TPS, ...tps });
     } catch (e) {
       result = `Error: ${e.message}`;
     }
